# backend-MyKP/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyKP API Documentation",
        "description": "API documentation for the MyKP backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bookmarks": {
            "get": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "List the authenticated user's saved (bookmarked) activities, newest-first by save time",
                "operationId": "1fb0d12ef5da5cdfdcc709532c40d31f",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ActivityID as a string",
                                                "type": "string",
                                                "example": "1"
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Seminar Bela Negara & Anti Narkoba 2026"
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "http://10.0.0.5:8000/images/seminarAntiNarkoba.jpeg",
                                                "nullable": true
                                            },
                                            "organizer": {
                                                "type": "string",
                                                "example": "BMA"
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "Talkshow (Wajib BMA)"
                                            },
                                            "points": {
                                                "type": "integer",
                                                "example": 6
                                            },
                                            "date": {
                                                "type": "string",
                                                "example": "Friday, 29 May 2026"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Save (bookmark) an activity for the authenticated user",
                "operationId": "ea5e1d78c03dc599b884aefaf8705143",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "activity_id"
                                ],
                                "properties": {
                                    "activity_id": {
                                        "description": "ActivityID to bookmark",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created — new bookmark inserted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "saved": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "activity_id": {
                                            "type": "string",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK — already bookmarked, no change",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "saved": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "activity_id": {
                                            "type": "string",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error — missing or non-existent activity_id"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bookmarks/{activity_id}": {
            "get": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Check whether the authenticated user has bookmarked a specific activity",
                "operationId": "5a64ddd62e43c6da4a2824cf87653918",
                "parameters": [
                    {
                        "name": "activity_id",
                        "in": "path",
                        "description": "ActivityID to check",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "saved": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "activity_id": {
                                            "type": "string",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Remove a bookmarked activity for the authenticated user",
                "operationId": "039d6cf7ab463474a1d4b11cb50452d4",
                "parameters": [
                    {
                        "name": "activity_id",
                        "in": "path",
                        "description": "ActivityID to un-bookmark",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK — bookmark removed (idempotent: returns 200 even if it wasn't bookmarked)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "saved": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "activity_id": {
                                            "type": "string",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum API Token",
                "bearerFormat": "string",
                "scheme": "bearer"
            }
        }
    },
...
}